requirements.txt
Flask==1.1.2
Flask-JWT-Extended==4.2.1
Flask-SQLAlchemy==2.5.1
passlib==1.7.2
gunicorn==20.0.4
paginate-sqlalchemy==0.3.1
